+++ src/components/BulkEditPane/BulkEditListResult/index.js
@@ -1,0 +1,1 @@
+export { default as BulkEditListResult } from './BulkEditListResult';