+++ src/components/BulkEditPane/BulkEditListSidebar/IdentifierTab/IdentifierTab.js
@@ -47,7 +47,7 @@ export const IdentifierTab = () => {
     identifier
   } = useSearchParams();
 
-  const [isDropZoneActive, setDropZoneActive] = useState(false);
+  const [isDropZoneActive, setIsDropZoneActive] = useState(false);
   const { fileUpload, isLoading } = useUpload();
   const { bulkOperationStart } = useBulkOperationStart();
 
@@ -63,7 +63,7 @@ export const IdentifierTab = () => {
 
   const [recordType] = activeFilters[IDENTIFIER_FILTERS.CAPABILITIES] || [];
   const [recordIdentifier] = activeFilters[IDENTIFIER_FILTERS.IDENTIFIER] || [];
-  const isDropZoneDisabled = isFileUploaded || !recordIdentifier || initialFileName;
+  const isDropZoneDisabled = isFileUploaded || !recordIdentifier || !!initialFileName;
   const capabilitiesFilterOptions = getCapabilityOptions(criteria, permissions);
 
   const {
@@ -124,11 +124,11 @@ export const IdentifierTab = () => {
   };
 
   const handleDragEnter = () => {
-    setDropZoneActive(true);
+    setIsDropZoneActive(true);
   };
 
   const handleDragLeave = () => {
-    setDropZoneActive(false);
+    setIsDropZoneActive(false);
   };
 
   const uploadFileFlow = async (fileToUpload) => {
@@ -148,8 +148,8 @@ export const IdentifierTab = () => {
       if (status === JOB_STATUSES.FAILED) throw Error();
 
       history.replace({
-        pathname: `/bulk-edit/${id}/progress`,
-        search: buildSearch({ fileName: fileToUpload.name }, location.search),
+        pathname: `/bulk-edit/${id}/preview`,
+        search: buildSearch({ fileName: fileToUpload.name, progress: CRITERIA.IDENTIFIER }, location.search),
       });
 
       setIsFileUploaded(true);
@@ -173,7 +173,7 @@ export const IdentifierTab = () => {
 
     await uploadFileFlow(fileToUpload);
 
-    setDropZoneActive(false);
+    setIsDropZoneActive(false);
   };
 
   const uploaderSubTitle = useMemo(() => {