+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * [UIBULKED-415](https://issues.folio.org/browse/UIBULKED-415) Query Tab - Support for Instance and Holdings Record Types.
 * [UIBULKED-434](https://folio-org.atlassian.net/browse/UIBULKED-434) Electronic access table is resized incorrectly when to go through pagination and back.
 * [UIBULKED-451](https://issues.folio.org/browse/UIBULKED-451) Set Maximum Limit for Bulk Edit.
+* [UIBULKED-416](https://issues.folio.org/browse/UIBULKED-416) Refactoring of bulk-edit preview components.
 
 ## [4.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v4.0.0) (2023-10-12)